backend/pyproject.toml
[project]
name = "credit-card-fraud-detection-dashboard"
version = "0.1.0"
description = "Fraud detection System"
requires-python = "==3.11.13"

dependencies = [
  "annotated-types==0.7.0",
  "anyio==4.10.0",
  "bleach==6.2.0",
  "certifi==2025.8.3",
  "charset-normalizer==3.4.3",
  "click==8.2.1",
  "dnspython==2.7.0",
  "email-validator==2.2.0",
  "fastapi==0.116.1",
  "fastapi-cli==0.0.8",
  "fastapi-cloud-cli==0.1.5",
  "h11==0.16.0",
  "httpcore==1.0.9",
  "httptools==0.6.4",
  "httpx==0.28.1",
  "idna==3.10",
  "iniconfig==2.1.0",
  "jinja2==3.1.6",
  "joblib==1.5.1",
  "kaggle==1.7.4.5",
  "markdown-it-py==4.0.0",
  "markupsafe==3.0.2",
  "mdurl==0.1.2",
  "numpy==1.26.4",
  "packaging==25.0",
  "pandas==2.3.2",
  "pluggy==1.6.0",
  "protobuf==6.32.0",
  "psycopg2-binary==2.9.10",
  "pydantic==2.11.7",
  "pydantic-core==2.33.2",
  "pygments==2.19.2",
  "python-dateutil==2.9.0.post0",
  "python-dotenv==1.1.1",
  "python-multipart==0.0.20",
  "python-slugify==8.0.4",
  "pytz==2025.2",
  "pyyaml==6.0.2",
  "requests==2.32.5",
  "rich==14.1.0",
  "rich-toolkit==0.15.0",
  "rignore==0.6.4",
  "sentry-sdk==2.35.0",
  "setuptools==80.9.0",
  "shellingham==1.5.4",
  "six==1.17.0",
  "sniffio==1.3.1",
  "sqlalchemy==2.0.43",
  "starlette==0.47.2",
  "text-unidecode==1.3",
  "tqdm==4.67.1",
  "typer==0.16.1",
  "typing-extensions==4.14.1",
  "typing-inspection==0.4.1",
  "tzdata==2025.2",
  "urllib3==2.5.0",
  "uvicorn==0.35.0",
  "uvloop==0.21.0",
  "watchfiles==1.1.0",
  "webencodings==0.5.1",
  "websockets==15.0.1",
  "scipy==1.16.1",
  "xgboost==2.0.3",
  "scikit-learn==1.2.2",
  "threadpoolctl==3.6.0",
  "docker==7.1.0",
  "testcontainers==4.12.0",
  "wrapt==1.17.3",
  "pytest==8.4.1",
  "pytest-asyncio==1.1.0",
  "prometheus-fastapi-instrumentator>=7.1.0",
  "asyncpg==0.30.0",
  "passlib[bcrypt]==1.7.4",
  "pyjwt==2.10.1",
  "websockets==15.0.1",
  "aiohttp==3.11.10",
  "google-auth==2.35.0",
  "colorama==0.4.6"
]
